refactor: RandomNumberGenerator static 삭제
* RandomNumberGenerator의 생명주기로 인해 랜덤한 숫자가 생성되지 않아 static 제거
