+++ src/main/java/racingcar/service/RandomNumberGenerator.java
@@ -1,9 +1,9 @@
-package racingcar.util;
+package racingcar.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class RandomNumberGenerator {
-	public static int generateRandomNumber(){
+	public int generateRandomNumber(){
 		return Randoms.pickNumberInRange(0, 9);
 	}
 }
